pyproject.toml
[project]
name = "aieng"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "datasketch>=1.6.5",
    "pyprobables>=0.6.1",
]
